# swagger.yaml
openapi: 3.0.3
info:
  title: Dice Engage — Omnichannel SaaS Platform API
  description: |
    Production-ready, multi-tenant omnichannel marketing SaaS API.
    Supports Email, SMS, WhatsApp, Mobile Push, Web Push, RCS, and Webhooks
    with tenant isolation, billing, analytics, and admin management.
  version: 1.0.0
  contact:
    name: Dice Engage Team
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.diceengage.com
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      name: Authorization
      in: header
    TenantContext:
      type: apiKey
      name: x-tenant-id
      in: header
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    UsageSummary:
      type: object
      properties:
        tenantId:
          type: string
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        emailCount:
          type: integer
        smsCount:
          type: integer
        whatsappCount:
          type: integer
        pushCount:
          type: integer
        webhookCount:
          type: integer
        totalMessages:
          type: integer
        totalCostMicros:
          type: integer
          description: Total cost in micros (1,000,000 micros = $1 USD)
        currency:
          type: string
    PlanLimitCheck:
      type: object
      properties:
        allowed:
          type: boolean
        currentUsage:
          type: integer
        limit:
          type: integer
        plan:
          type: string
          enum: [Starter, Growth, Enterprise]
    ChannelBreakdown:
      type: object
      properties:
        channel:
          type: string
        count:
          type: integer
        costMicros:
          type: integer
...
security:
  - BearerAuth: []
paths:
...
  /api/admin/billing/usage/{tenantId}:
    get:
      tags: [Billing]
      summary: Get current billing period usage for a tenant
      operationId: getBillingUsage
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Current billing period usage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageSummary"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/billing/limits/{tenantId}:
    get:
      tags: [Billing]
      summary: Check plan limits for a tenant
      operationId: checkPlanLimits
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: channel
          in: query
          schema:
            type: string
            enum: [Email, Sms, WhatsApp, MobilePush, WebPush, Webhook, RCS]
          description: Channel to check limits for (defaults to Email)
      responses:
        "200":
          description: Plan limit check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanLimitCheck"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/billing/analytics/{tenantId}:
    get:
      tags: [Billing]
      summary: Get message analytics with channel breakdown
      operationId: getMessageAnalytics
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
          description: Start date (defaults to start of current month)
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
          description: End date (defaults to now)
        - name: workspaceId
          in: query
          schema:
            type: string
            format: uuid
          description: Optional workspace filter
      responses:
        "200":
          description: Channel breakdown analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  channels:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChannelBreakdown"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ═══════════════════════════════════════════════════════════════════
  # ADMIN SUPER PANEL — /api/admin/panel
  # ═══════════════════════════════════════════════════════════════════
...